requirements.txt
async-generator==1.10
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
oauthlib==3.1.1
outcome==1.1.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
wsproto==1.0.0
